WEB-3834: Updating the CLI to support books and videos
